LayoutTests/http/tests/history/cross-origin-replace-history-object-child-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/history/resources/cross-origin-replaces-history-object-child-iframe.html from frame with URL http://127.0.0.1:8000/history/cross-origin-replace-history-object-child.html. Domains, protocols and ports must match.

ALERT: Child window's history object before attempt to clear: [object History]
ALERT: About to shadow child window's history object: [object History]
ALERT: Shadowed child window's history object: 
CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/history/resources/cross-origin-replaces-history-object-child-iframe.html from frame with URL http://127.0.0.1:8000/history/cross-origin-replace-history-object-child.html. Domains, protocols and ports must match.

ALERT: Child window's history object after attempt to clear: [object History]
